+++ packages/webblocks-cms/src/Http/Controllers/Admin/SystemSettingsController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\RedirectResponse;
 use Illuminate\Routing\Controller;
 use Illuminate\View\View;
 use WebBlocks\Cms\Http\Requests\Admin\SystemSettingsRequest;
+use WebBlocks\Cms\Support\Mail\CmsMailSettingsResolver;
 use WebBlocks\Cms\Support\System\InstalledVersionStore;
 use WebBlocks\Cms\Support\System\SystemSettings;
 use WebBlocks\Cms\WebBlocksCmsServiceProvider;
@@ -15,6 +16,7 @@ class SystemSettingsController extends Controller
   public function __construct(
     private readonly SystemSettings $systemSettings,
     private readonly InstalledVersionStore $installedVersionStore,
+    private readonly CmsMailSettingsResolver $mailSettingsResolver,
   ) {}
 
   public function edit(): View
@@ -30,7 +32,20 @@ class SystemSettingsController extends Controller
         'timezone' => old('timezone', $this->systemSettings->timezone()),
         'admin_listing_per_page' => old('admin_listing_per_page', $this->systemSettings->adminListingPerPage()),
         'visitor_consent_banner_enabled' => old('visitor_consent_banner_enabled', $this->systemSettings->visitorConsentBannerEnabled()),
+        'cms_mail_mode' => old('cms_mail_mode', $this->systemSettings->cmsMailSettings()['mode']),
+        'cms_mail_mailer' => old('cms_mail_mailer', $this->systemSettings->cmsMailSettings()['mailer']),
+        'cms_mail_host' => old('cms_mail_host', $this->systemSettings->cmsMailSettings()['host']),
+        'cms_mail_port' => old('cms_mail_port', $this->systemSettings->cmsMailSettings()['port']),
+        'cms_mail_encryption' => old('cms_mail_encryption', $this->systemSettings->cmsMailSettings()['encryption']),
+        'cms_mail_username' => old('cms_mail_username', $this->systemSettings->cmsMailSettings()['username']),
+        'cms_mail_from_address' => old('cms_mail_from_address', $this->systemSettings->cmsMailSettings()['from_address']),
+        'cms_mail_from_name' => old('cms_mail_from_name', $this->systemSettings->cmsMailSettings()['from_name']),
+        'cms_mail_reply_to_address' => old('cms_mail_reply_to_address', $this->systemSettings->cmsMailSettings()['reply_to_address']),
+        'cms_mail_timeout' => old('cms_mail_timeout', $this->systemSettings->cmsMailSettings()['timeout']),
+        'cms_mail_password_configured' => $this->systemSettings->cmsMailPasswordConfigured(),
       ],
+      'mailDiagnostics' => $this->mailSettingsResolver->diagnostics(),
+      'cmsMailMailerOptions' => array_combine(CmsMailSettingsResolver::SUPPORTED_MAILERS, CmsMailSettingsResolver::SUPPORTED_MAILERS),
       'localeOptions' => $this->systemSettings->enabledLocaleOptions(),
       'timezoneOptions' => $this->systemSettings->timezoneOptions(),
       'installedVersionDisplay' => $this->installedVersionStore->displayVersion(),