+++ packages/webblocks-cms/src/Support/Mail/CmsMailConfigurationException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace WebBlocks\Cms\Support\Mail;
+
+use RuntimeException;
+
+class CmsMailConfigurationException extends RuntimeException {}